requirements.txt
django==3.2
djangorestframework==3.12
djangorestframework-simplejwt==5.0.0
djoser==2.0
six==1.16.0
psycopg2-binary==2.9.3
flake8==4.0.1
flake8-commas==2.1.0
flake8-import-order==0.18.1
pytest==7.0.0
pytest-django==4.5.2
pytest-cov==3.0.0
celery==5.2.3
redis==4.1.4